task-8-redis-image-resize/requirements.txt
flask>=1.0.2
gunicorn>=19.7.1
flask-restplus>=0.12.1
flake8>=3.7
black>=18.9b0
pylint>=2.2
pytest>=4.4.0
pytest-cov>=2.6
pytest-mock>=1.10
pytest-flask>=0.14.0
redis>=3.2.1
rq>=1.0
Pillow>=6.0.0 